CACluster_Points
Excel file (ClustersFinal30.xls) contains the latitude/longitude positions of the cluster points. Clust35 is the point locations for 35 climate distinct regions and Clust30 is the point locations for 30 distinct climate regions.

Shapefile (CACLusters30.zip) contains the file for the point locations of the 30 distinct climate regions. This file was used in subsequent analysis to assign 30 climate regions to ZIP-codes across CA.